// blog-server/src/common/enums/error-code.enum.ts
export enum ErrorCode {
  FORBIDDEN_RESOURCE = 'FORBIDDEN_RESOURCE',
  ACCOUNT_EXISTED = 'ACCOUNT_EXISTED',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  TOKEN_EXPIRES = 'TOKEN_EXPIRES',
  TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
  // ROLES ERROR
  ROLE_EXISTED = 'ROLE_EXISTED',
  // SCOPES ERROR
  SCOPE_EXISTED = 'SCOPE_EXISTED',
  // CATEGORY ERROR
  CATEGORY_EXISTED = 'CATEGORY_EXISTED',
}